Explicitly require go get arduino-cli
I think until go modules are the default for development
(coming in go1.13) we have to explicitly require
a go get github.com/arduino/arduino-cli as it can't be
imported directly. Only packages can be imported not
programs. And not being able to import means it won't be
built as a dep when running
go get github.com/robgonnella/ardie

Also removes sketches directory as this is just a command
line tool now not a template