Use rpc client for better parsing and control
Rather than trying to parse the output text from the command-line
tool, we temporarily start the grpc server locally and make requests
so we can get typed responses. This makes it easier to get and parse
the raw board list as well as monitor the output of other calls like
"compile" and "upload".