Watch for "REMOVE" events when watching for file changes
Editors that make atomic updates will create temporary files and when
ready to "write" will rename these tmp files as the target file. This
triggers a "REMOVE" evt as the final evt in the sequence. By watching
for this event we can check if the target file still exists and then
perform our updates as expected.

Issue #55